requirements.txt
alpaca==1.0.0
alpaca-py==0.43.2
annotated-types==0.7.0
apify_client==2.5.0
apify_shared==2.2.0
appnope==0.1.4
asttokens==3.0.1
aurora-dsql-python-connector==0.2.6
boto3==1.42.65
botocore==1.42.65
certifi==2026.2.25
charset-normalizer==3.4.5
colorama==0.4.6
comm==0.2.3
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
homebrew==0.0.0.1
idna==3.11
impit==0.12.0
ipykernel==7.2.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jmespath==1.1.0
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
matplotlib-inline==0.2.1
more-itertools==10.8.0
msgpack==1.1.2
nest-asyncio==1.6.0
numpy==2.4.3
packaging==26.0
pandas==3.0.1
parso==0.8.6
pexpect==4.9.0
platformdirs==4.9.4
prompt_toolkit==3.0.52
psutil==7.2.2
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg-pool==3.3.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
pyzmq==27.1.0
requests==2.32.5
s3transfer==0.16.0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
sseclient-py==1.9.0
stack-data==0.6.3
threadpoolctl==3.6.0
tornado==6.5.5
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wcwidth==0.6.0
websockets==16.0
xgboost==3.2.0
